:bug: Fix refactoring errors
